+++ src/constants/index.js
@@ -4,3 +4,5 @@ exports.PAGE_PER_SCREEN = 4;
 exports.POST = 'post';
 exports.PORTFOLIO = 'portfolio';
 exports.RESUME = 'resume';
+
+exports.SITE_URL = 'https://wonism.github.io';