+++ src/store/app/initialState.js
@@ -2,7 +2,6 @@ const isProduction = JSON.stringify(process.env.NODE_ENV) === 'production';
 
 export default {
   isProduction,
-  isClipboardSupported: false,
   searchKeyword: '',
   portfolios: [],
   postInformations: [],