+++ src/store/app/initializer.js
@@ -1,9 +1,1 @@
-import { INIT_CLIPBOARD_SUPPORT } from './actionTypes';
-
-export default (store) => {
-  store.dispatch({
-    type: INIT_CLIPBOARD_SUPPORT,
-  });
-
-  return store;
-};
+export default store => store;