+++ src/store/app/reducers.js
@@ -1,7 +1,5 @@
 import fp from 'lodash/fp';
 import {
-  SUPPORT_CLIPBOARD,
-  DO_NOT_SUPPORT_CLIPBOARD,
   INPUT_KEYWORD,
   OPEN_MENU,
   CLOSE_MENU,
@@ -10,8 +8,6 @@ import {
 } from './actionTypes';
 
 export default {
-  [SUPPORT_CLIPBOARD]: fp.set('isClipboardSupported', true),
-  [DO_NOT_SUPPORT_CLIPBOARD]: fp.set('isClipboardSupported', false),
   [INPUT_KEYWORD]: (state, { input }) => fp.set('searchKeyword', input)(state),
   [OPEN_MENU]: fp.set('isMenuOpened', true),
   [CLOSE_MENU]: fp.set('isMenuOpened', false),