+++ src/templates/CategorizedList.jsx
@@ -56,7 +56,7 @@ const CategorizedList = ({
           const { images, tags, path } = frontmatter;
 
           return (
-            <Card key={path} path={path} images={images} tags={tags} {...frontmatter} hasTags={!fp.isEmpty(tags)} />
+            <Card key={path} path={path} images={images} tags={tags} {...frontmatter} />
           );
         }