+++ src/templates/List.jsx
@@ -34,7 +34,7 @@ const List = ({
           const { images, tags, path } = frontmatter;
 
           return (
-            <Card key={path} path={path} images={images} tags={tags} {...frontmatter} hasTags={!fp.isEmpty(tags)} />
+            <Card key={path} path={path} images={images} tags={tags} {...frontmatter} />
           );
         }