+++ src/templates/TaggedList.jsx
@@ -56,7 +56,7 @@ const TaggedList = ({
           const { images, tags, path } = frontmatter;
 
           return (
-            <Card key={path} path={path} tags={tags} images={images} {...frontmatter} hasTags={!fp.isEmpty(tags)} />
+            <Card key={path} path={path} tags={tags} images={images} {...frontmatter} />
           );
         }